working selector and splitting out formatting
